# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Helsa API
  version: 0.1.0
  description: >
    Helsa — Apple Health / fitness data processing backend API.
    Contract-first: this spec is the single source of truth (the backend, the web and the Apple clients all generate from it).
    A single-user system, on home infrastructure (see docs/decisions/ADR-0002).
    Access: /ingest and the read endpoints sit behind an mTLS client certificate, with a device token above it;
    the web dashboard is reachable over WireGuard only (docs/08-auth-hozzaferes.md).
    Every timestamp is ISO-8601 UTC; the daily/weekly aggregates take a tz parameter.
servers:
  - url: https://api.helsa.example.com/v1
    description: home VM, an mTLS client certificate is mandatory (the final hostname is TBD — docs/07 §10)
  - url: http://localhost:8080/v1
    description: local
security:
  - deviceToken: []
...
paths:
  /auth/refresh:
    post:
      tags: [auth]
      summary: Refresh the access token with a refresh token
      security: []
      responses:
        '200':
          description: A new session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        '401': { $ref: '#/components/responses/Problem' }
...
  /workouts/{id}:
    get:
      tags: [read]
      summary: Workout detail
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Workout
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workout' }
        '404': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    deviceToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        A long-lived device token (docs/08-auth-hozzaferes.md §7). It does NOT come from an interactive
        sign-in — an install-time or manual step issues it. This is the application-level factor; the
        transport-level factor is the mTLS client certificate that the reverse proxy verifies (the backend
        sees which device is speaking from the X-Client-Cert-CN header).
  responses:
    Problem:
      description: Error (RFC 9457)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
    Session:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string, description: "for the native clients; on the web it lives in an httpOnly cookie" }
        expires_in: { type: integer }
...
    Workout:
      type: object
      properties:
        id: { type: string, format: uuid }
        source_uuid: { type: string }
        activity_type: { type: string }
        started_at: { type: string, format: date-time }
        ended_at: { type: string, format: date-time }
        total_energy_kcal: { type: number }
        total_distance_m: { type: number }
        avg_heart_rate: { type: number }
        max_heart_rate: { type: number }
        metadata: { type: object, additionalProperties: true }
...