# backend/internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Achievement struct {
	UserID     pgtype.UUID        `json:"user_id"`
	ID         string             `json:"id"`
	Kind       string             `json:"kind"`
	Code       string             `json:"code"`
	Period     *string            `json:"period"`
	Value      *float64           `json:"value"`
	Unit       *string            `json:"unit"`
	Thresholds []int32            `json:"thresholds"`
	EarnedAt   pgtype.Timestamptz `json:"earned_at"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
	UpdatedAt  pgtype.Timestamptz `json:"updated_at"`
}

type ActivitySummary struct {
	UserID           pgtype.UUID        `json:"user_id"`
	Day              pgtype.Date        `json:"day"`
	ActiveEnergy     *float64           `json:"active_energy"`
	ActiveEnergyGoal *float64           `json:"active_energy_goal"`
	ExerciseMinutes  *float64           `json:"exercise_minutes"`
	ExerciseGoal     *float64           `json:"exercise_goal"`
	StandHours       *float64           `json:"stand_hours"`
	StandGoal        *float64           `json:"stand_goal"`
	IngestedAt       pgtype.Timestamptz `json:"ingested_at"`
}

type Device struct {
	ID         pgtype.UUID        `json:"id"`
	UserID     pgtype.UUID        `json:"user_id"`
	Platform   string             `json:"platform"`
	Model      *string            `json:"model"`
	Name       *string            `json:"name"`
	TimeZone   *string            `json:"time_zone"`
	LastSeenAt pgtype.Timestamptz `json:"last_seen_at"`
}

type Goal struct {
	ID          pgtype.UUID        `json:"id"`
	UserID      pgtype.UUID        `json:"user_id"`
	Metric      string             `json:"metric"`
	TargetValue float64            `json:"target_value"`
	Unit        string             `json:"unit"`
	Source      string             `json:"source"`
	HkValue     *float64           `json:"hk_value"`
	UpdatedAt   pgtype.Timestamptz `json:"updated_at"`
}

type PushToken struct {
	ID          pgtype.UUID        `json:"id"`
	UserID      pgtype.UUID        `json:"user_id"`
	DeviceID    pgtype.UUID        `json:"device_id"`
	Token       string             `json:"token"`
	Platform    string             `json:"platform"`
	Environment string             `json:"environment"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
	UpdatedAt   pgtype.Timestamptz `json:"updated_at"`
}

type Sample struct {
	UserID            pgtype.UUID        `json:"user_id"`
	Ts                pgtype.Timestamptz `json:"ts"`
	DataType          string             `json:"data_type"`
	Value             *float64           `json:"value"`
	Unit              *string            `json:"unit"`
	WorkoutID         pgtype.UUID        `json:"workout_id"`
	SourceUuid        string             `json:"source_uuid"`
	SourceDevice      *string            `json:"source_device"`
	SourceBundle      *string            `json:"source_bundle"`
	IngestedAt        pgtype.Timestamptz `json:"ingested_at"`
	WorkoutSourceUuid *string            `json:"workout_source_uuid"`
}

type SleepSegment struct {
	ID         pgtype.UUID        `json:"id"`
	UserID     pgtype.UUID        `json:"user_id"`
	SourceUuid string             `json:"source_uuid"`
	StartedAt  pgtype.Timestamptz `json:"started_at"`
	EndedAt    pgtype.Timestamptz `json:"ended_at"`
	Stage      string             `json:"stage"`
	IngestedAt pgtype.Timestamptz `json:"ingested_at"`
}

type SyncState struct {
	ID        pgtype.UUID        `json:"id"`
	UserID    pgtype.UUID        `json:"user_id"`
	DataType  string             `json:"data_type"`
	Anchor    *string            `json:"anchor"`
	UpdatedAt pgtype.Timestamptz `json:"updated_at"`
}

type User struct {
	ID          pgtype.UUID        `json:"id"`
	AppleSub    string             `json:"apple_sub"`
	Email       *string            `json:"email"`
	DisplayName *string            `json:"display_name"`
	TimeZone    string             `json:"time_zone"`
	Locale      string             `json:"locale"`
	UnitSystem  string             `json:"unit_system"`
	NotifPrefs  []byte             `json:"notif_prefs"`
	ConsentAt   pgtype.Timestamptz `json:"consent_at"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
}

type Workout struct {
	ID              pgtype.UUID        `json:"id"`
	UserID          pgtype.UUID        `json:"user_id"`
	SourceUuid      string             `json:"source_uuid"`
	ActivityType    string             `json:"activity_type"`
	StartedAt       pgtype.Timestamptz `json:"started_at"`
	EndedAt         pgtype.Timestamptz `json:"ended_at"`
	TotalEnergyKcal *float64           `json:"total_energy_kcal"`
	TotalDistanceM  *float64           `json:"total_distance_m"`
	Metadata        []byte             `json:"metadata"`
}

type WorkoutRoutePoint struct {
	WorkoutID pgtype.UUID        `json:"workout_id"`
	Seq       int32              `json:"seq"`
	Ts        pgtype.Timestamptz `json:"ts"`
	Lat       float64            `json:"lat"`
	Lon       float64            `json:"lon"`
	AltitudeM *float64           `json:"altitude_m"`
	SpeedMps  *float64           `json:"speed_mps"`
	AccuracyM *float64           `json:"accuracy_m"`
}